Rename algorithms/map/longest_common_subsequence.py to longest_common_substring.py
The file under algorithms/map/ actually implements longest common
*substring* (contiguous), not longest common subsequence. The module
docstring, the function name (`max_common_sub_string`), and the Wikipedia
reference URL inside the file all already say "substring" — only the
filename was wrong. The true longest common subsequence lives in
algorithms/dynamic_programming/longest_common_subsequence.py.

Updates the single import site in algorithms/map/__init__.py. The
exported symbol name (`max_common_sub_string`) is unchanged, so no
test or downstream caller needs to change.

Closes #2747.
