+++ algorithms/map/__init__.py
@@ -9,7 +9,7 @@ from algorithms.data_structures.separate_chaining_hash_table import (
 
 from .is_anagram import is_anagram
 from .is_isomorphic import is_isomorphic
-from .longest_common_subsequence import max_common_sub_string
+from .longest_common_substring import max_common_sub_string
 from .longest_palindromic_subsequence import longest_palindromic_subsequence
 from .randomized_set import RandomizedSet
 from .valid_sudoku import is_valid_sudoku
