+++ algorithms/map/longest_common_substring.py
